Improve structure, load files from subdirs
