+++ src/combine.ts
@@ -1,0 +1,3 @@
+export function combine<T>(left: T[] | undefined, right: T[]): T[] {
+  return [...(left || []), ...right];
+}
