+++ test/module/subdir/subdir.controller.ts
