+++ test/module/subdir/subdir.provider.ts
